refactor(frontend): extract the new-session modal cluster from main.ts (#312)
The largest remaining surface: the create-session modal and its four tabs
(repo #188, advanced #191, env #186, ports #190) plus the save-as-template
flow (#195) and the templates page (#195) — ~1700 lines — into newSession.ts.

Surprisingly low coupling for its size: it reads only `sessions` from core,
owns all its own state (envRows/portRows/editingTemplateId/newSessionOpener/
templatesOpener), and imports four core functions (openSession,
refreshSessions, renderSessionList, showToast). Reverse deps exported back to
main.ts: the three modal close functions (Escape handler) +
applyResourceCapsToForm + openBootstrapLogModal. Same TDZ-safe pattern: DOM
re-queried locally; the three shared modal consts (newSessionModal,
saveTemplateModal, templatesModal) re-queried in both modules for main's
modal guard.

main.ts: 3478 → 1766 (and 5229 → 1766 over the whole #312 split — −66%).
tsc + Biome clean, vite build green, frontend tests pass (66).
